+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-master
     * HOTFIX      #2864 [MediaBundle]         Return 404 http code for not existing media
+    * HOTFIX      #2865 [ContentBundle]       Fixed button text for unpublish dialog
     * HOTFIX      #2804 [ListBuilder]         Fixed ids-query in doctrine-list-builder
     * HOTFIX      #2839 [WebsiteBundle]       Include port in URLs